level_three/dashboard-app/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.3
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
